# Remove timing difference between 90° and 270° rotations (in the tiling Rust code)