@surma 2019-02-21
Stellar work. I’m gonna merge this! I am as surprised as you are that these two loops were show such contrasting performance characteristics.

cc @ThomasTheDane and @dschuff